Support symlinks to identity files
With this change it is sufficient to unlock only one store when another
store uses the same identity (via a symlink), because the senior-agent
now uses the canonicalised path of the identity file

Example setup:
  store1/.identity.age
  store2/.identity.age -> ../store1/.identity.age
  OR
  store1/.identity.age -> /path/to/identityfile
  store2/.identity.age -> /path/to/identityfile

After `senior -s store1 show name1` you are asked for the passphrase of
the *canonicalised* identity file. The senior-agent stores the
passphrase of this *canonicalised* identity file.
When you then want to view a passphrase from store2 via
`senior -s store2 show name2` you are not asked for a passphrase
anymore, because the *canonicalised* identity file of store2 is the same
as for store1.