More security for senior change-passphrase
A backup of the identity is written to a temporary directory and the
user is informed.
Additionally the agent now supports deleting a key. This is done before
a passphrase change. Otherwise the user might try `senior show ...`, and
get the identity from the agent instead of the new file. The user might
think the passphrase change was successful even though their new
identity file might be broken.